moved widgest into header/source files

+++ include/Viewer/Core/Camera.h
@@ -147,7 +147,7 @@ namespace VkRender {
             m_Fov = fov;
             m_Znear = zNear;
             m_Zfar = zFar;
-            float focal_length = 1.0f / tan(glm::radians(m_Fov)* 0.5f) ;
+            float focal_length = 1.0f / tanf(glm::radians(m_Fov)* 0.5f) ;
             float x = focal_length / aspect;
             float y = -focal_length;
             float A = -m_Zfar / (m_Zfar - m_Znear);
